// apps/portal-api/src/model/kanel/public/DocumentChildren.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { DocumentId } from './Document';

/** Represents the table public.Document_Children */
export default interface DocumentChildren {
  parent_document_id: DocumentId;

  child_document_id: DocumentId;
}

/** Represents the initializer for the table public.Document_Children */
export interface DocumentChildrenInitializer {
  parent_document_id: DocumentId;

  child_document_id: DocumentId;
}

/** Represents the mutator for the table public.Document_Children */
export interface DocumentChildrenMutator {
  parent_document_id?: DocumentId;

  child_document_id?: DocumentId;
}